extracted tests from suite
